quiz-server/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-jose==3.5.0
python-multipart==0.0.21
python-dotenv==1.0.1
pydantic==2.12.5
